Move mypyc.genclass to mypyc.irbuild.classdef
