+++ mypyc/genopsvisitor.py
@@ -21,7 +21,7 @@ from mypy.nodes import (
 
 from mypyc.ops import Value
 from mypyc.genops import IRVisitor, IRBuilder, UnsupportedException
-from mypyc.genclass import transform_class_def
+from mypyc.irbuild.classdef import transform_class_def
 from mypyc.irbuild.function import (
     transform_func_def,
     transform_overloaded_func_def,
