+++ mypyc/irbuild/classdef.py
